% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_hier.R
\name{age_hier}
\alias{age_hier}
\title{Creating GSS Age Hierarchy}
\usage{
age_hier(cutoff_age = 80)
}
\arguments{
\item{cutoff_age}{the maximum age to be used in the age groups (e.g. 65, 80, 100)}
}
\value{
Output is a sdc hierarchy of the age groups used by GSS.
}
\description{
Creating GSS Age Hierarchy
}
\examples{
age_hier()
}
