% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_median.R
\name{calc_mean}
\alias{calc_mean}
\title{Compute the mean and meadian age of the data}
\usage{
calc_mean(combination, data, variable, output_var, weight_var)
}
\arguments{
\item{combination}{variabes names to consider in the data frame.}

\item{data}{an object (dataframe).}

\item{variable}{focus variable a numeric vector.}

\item{output_var}{output variable.}

\item{weight_var}{weight variable (must be numeric).}
}
\value{
data frame
}
\description{
Compute the mean and meadian age of the data
}
