% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_age_data.R
\name{clean_age_data}
\alias{clean_age_data}
\title{Clean Age Variable of Data Create Age Hierarchy}
\usage{
clean_age_data(
  data,
  age_var = "age",
  min_age = 0,
  max_age = 100,
  max_age_type = "cutoff",
  age_group = FALSE,
  print_hier = FALSE
)
}
\arguments{
\item{data}{Data with the age variable that you want to clean and create the hierarchy for}

\item{age_var}{Variable name of the age variable (e.g. "age")}

\item{min_age}{Minimum age the data should start at}

\item{max_age}{Maximum age the data should stop at}

\item{max_age_type}{cutoff OR max: A cutoff labels values above the maximum as 'max+' (e.g., 80+), while a max drops all the values above the max value.}

\item{age_group}{TRUE or FALSE, whether age should be replaced by groups, or single ages should be preserved}

\item{print_hier}{TRUE or FALSE, whether the resulting hierarchy should be printed}
}
\value{
List with cleaned data and age hierarchy
}
\description{
Clean Age Variable of Data Create Age Hierarchy
}
