% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_hier.R
\name{geo_hier}
\alias{geo_hier}
\title{Creating Ghana's Geographic Hierarchy}
\usage{
geo_hier()
}
\value{
Output is a sdc hierarchy of the regions, districts and subdistricts of Ghana.
}
\description{
Creating Ghana's Geographic Hierarchy
}
\examples{
geo_hier()
}
