% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_hierarchie.R
\name{make_sub_hier}
\alias{make_sub_hier}
\title{Create Sub-Hierarchy}
\usage{
make_sub_hier(data, var_total, var_main, sub_var, main_groups)
}
\arguments{
\item{data}{A data frame containing the data to be used for creating the hierarchy.}

\item{var_total}{A string specifying the root node for the hierarchy.}

\item{var_main}{A string specifying the column name in `data` containing the main categories (roots).}

\item{sub_var}{A string specifying the column name in `data` containing the sub-categories (children).}

\item{main_groups}{A vector specifying the categories under `var_main` for which child levels should be created.}
}
\value{
A hierarchical structure with the specified root and child nodes.
}
\description{
This function creates a hierarchical structure using the `sdcHierarchies` package. It dynamically adds child nodes for specified parent nodes based on the provided data.
}
