% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_px_converter.R
\name{old_process_data}
\alias{old_process_data}
\title{Old function that both processes data and converts it to a px file}
\usage{
old_process_data(
  data,
  var_names,
  var_totals,
  var_display_names,
  weight_var,
  save_location,
  min_cells = 5,
  rate_element = NULL,
  numericVar = NULL,
  add_hier = NULL,
  summary_type = "freq",
  ...
)
}
\arguments{
\item{data}{an object (dataframe)}

\item{var_names}{disaggregation variables}

\item{var_totals}{character for totals in all disaggregates.}

\item{var_display_names}{variable names to replace the disaggregation variables.}

\item{weight_var}{weight variable (must be numeric).}

\item{save_location}{the directory to save the returned output.}

\item{min_cells}{minimum frequency for suppression (default is 5)}

\item{rate_element}{focus element to retain under single_rate summary_type}

\item{numericVar}{a numeric or character vector defining the column indices or variable names of additional numeric variables with respect to data.}

\item{add_hier}{update the list of hierarchie created by this function with a list of different hierarchies}

\item{summary_type}{type of summary to make. Only accepts:
\itemize{
  \item **freq:** frequencies or counts
  \item **percent:** computes percenatages for dimensions.
  \item **numFreq:** computes frequency of numeric variable by dimensions. Eg. Number of livestock raised by households/individuals.
  \item **numSum:** computes sum of numeric variable by dimensions.
  \item **single_rate:** computes rates for dimensions. eg. unemployment rate without the inverse (employment).
}}

\item{...}{additional variables for as.px.data.frame}
}
\value{
pxfile
}
\description{
Old function that both processes data and converts it to a px file
}
