% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_converter.R
\name{process_data}
\alias{process_data}
\title{Convert data frame to px file}
\usage{
process_data(
  data,
  var_names,
  var_totals,
  var_display_names,
  weight_var,
  title,
  matrix,
  u_of_m = "percentage of population",
  subject_code,
  subject_area,
  content,
  showdecimals,
  decimals,
  language,
  source,
  note,
  save_location,
  min_cells = 5,
  summary_type = "freq",
  rate_element = NULL,
  numericVar = NULL
)
}
\arguments{
\item{data}{an object (dataframe)}

\item{var_names}{disaggregation variables}

\item{var_totals}{character for totals in all disaggregates.}

\item{var_display_names}{variable names to replace the disaggregation variables.}

\item{weight_var}{weight variable (must be numeric).}

\item{title}{title of the statistical table being generated.}

\item{matrix}{summary of the table or source.}

\item{u_of_m}{unit of measure.}

\item{subject_code}{subject code.}

\item{subject_area}{subject area.}

\item{content}{similar to title.}

\item{showdecimals}{numeric value showing the decimal place.}

\item{decimals}{numeric value showing the decimal place.}

\item{language}{en.}

\item{source}{source of data eg. GDHS and GLSS.}

\item{note}{notes.}

\item{save_location}{the directory to save the returned output.}

\item{min_cells}{minimum frequency for suppression (default is 5)}

\item{summary_type}{type of summary to make. Only accepts:
\itemize{
  \item **freq:** frequencies or counts
  \item **rate:** computes rates for dimensions.
}}

\item{numericVar}{a numeric or character vector defining the column indices or variable names of additional numeric variables with respect to data.}
}
\value{
px file
}
\description{
Convert data frame to px file
}
