% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_converter.R
\name{px_converter}
\alias{px_converter}
\title{Convert Processed Data into PX File}
\usage{
px_converter(
  data,
  title,
  matrix,
  units,
  contents,
  source,
  save_location,
  note = NULL,
  hier_variables = NULL,
  hierarchies = NULL,
  showdecimals = 0,
  decimals = 0,
  ...
)
}
\arguments{
\item{data}{Processed data that should be turned into a PX file}

\item{title}{Title of the table}

\item{matrix}{Matrix?}

\item{units}{Unit of values in the table}

\item{contents}{Content of the table}

\item{source}{Source of the data}

\item{note}{Notes for underneath the table}

\item{hier_variables}{Vector of variable names for which a hierarchy should be included}

\item{hierarchies}{List of hierarchies of previously listed variables (in same order)}

\item{showdecimals}{Number of decimals to be shown on the StatsBank}

\item{decimals}{Number of decimals to be seen when data is downloaded from the StatsBank}

\item{...}{Additional variables used to make px file}

\item{path}{Path and name under which the px file will be saved (e.g. output/table1.px)}
}
\value{
pxfile
}
\description{
Convert Processed Data into PX File
}
