% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_func.R
\name{toTitleCase}
\alias{toTitleCase}
\title{Set right format for px file titler}
\usage{
toTitleCase(s)
}
\arguments{
\item{s}{A string as title of the px file.}
}
\value{
string
}
\description{
Set right format for px file titler
}
