% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Data into Format Needed for PX File}
\usage{
process_data(
  data,
  var_names,
  var_totals,
  var_display_names,
  weight_var,
  add_hier = NULL,
  summary_type = "freq",
  rate_element = NULL,
  numericVar = NULL,
  min_cells = 5,
  suppress = FALSE
)
}
\arguments{
\item{data}{A dataframe with no missing values}

\item{var_names}{Disaggregation variables.}

\item{var_totals}{Dimension totals. This is the root for each dimension hierarchie.}

\item{var_display_names}{Specify new variable names to replace the disaggregation variables.}

\item{weight_var}{Weight variable (must be numeric).}

\item{add_hier}{Adds list of hierarchie}

\item{summary_type}{Type of summary to make. Only accepts:
\itemize{
  \item  \textbf{freq:} frequencies or counts
  \item **percent:** computes percentages for dimensions.
  \item **numFreq:** computes frequency of numeric variable by dimensions. Eg. Number of livestock raised by households/individuals.
  \item **numSum:** computes sum of numeric variable by dimensions.
  \item **single_rate:** computes rates for dimensions. eg. unemployment rate without the inverse (employment).
}}

\item{rate_element}{Focus element to retain under single_rate summary_type}

\item{numericVar}{a numeric or character vector defining the column indices or variable names of additional numeric variables with respect to data.}

\item{min_cells}{Minimum frequency for suppression (default is 5)}
}
\value{
data
}
\description{
Process Data into Format Needed for PX File
}
